spotify_data_hackers/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.5.0
pandas==1.5.3
google-cloud-storage==2.5.0
numpy==1.23.2
requests==2.28.1
pyarrow==14.0.1
